add some refactore

+++ src/internal/adapters/repository/redis/cache.repository.ts
@@ -1,6 +1,5 @@
 import * as cache from "../../cache";
 import { RedisClientType } from "redis";
-import { CONFIG } from "../../../../deploy";
 import { TYPES } from "../../../domain/types";
 import { inject, injectable } from "inversify";
 import { ICacheRepository } from "../../../ports";
@@ -10,43 +9,44 @@ import { Logger, PREFIXES } from "../../../application/utils/log";
 export class RedisCacheRepository implements ICacheRepository {
   constructor(
     @inject(TYPES.Redis) private cache: cache.Redis,
-    @inject(TYPES.APP_CONFIG) private cfg: CONFIG,
     @inject(TYPES.Logger) private logger: Logger
   ) {
     this.cache.client = this.cache.client as RedisClientType;
   }
 
   async get(key: string): Promise<string | null> {
+    if (!(await this.ping())) return null;
     return await this.cache.client.get(key);
   }
 
   async set(key: string, value: string): Promise<void> {
+    if (!(await this.ping())) return;
     await this.cache.client.set(key, value);
   }
 
-  async setRateWithTimeout(key: string, value: string): Promise<void> {
-    try {
-      await this.cache.client.set(key, value);
-      setTimeout(async () => {
-        await this.cache.client.del(key);
-      }, this.cfg.rateLimitTime * 60 * 1000);
-    } catch (e) {
-      this.logger.print(PREFIXES.GRPC_SERVER, e as Error, "setWithTimeout Error");
-    }
-  }
-
   async setWithTimeout(key: string, value: string, time: number): Promise<void> {
+    if (!(await this.ping())) return;
     try {
       await this.cache.client.set(key, value);
       setTimeout(async () => {
         await this.delete(key);
       }, time * 60 * 1000);
     } catch (e) {
-      this.logger.print(PREFIXES.GRPC_SERVER, e as Error, "setWithTimeout Error");
+      this.logger.print(PREFIXES.REDIS, e as Error, "setWithTimeout Error");
     }
   }
 
   async delete(key: string): Promise<void> {
+    if (!(await this.ping())) return;
     await this.cache.client.del(key);
+  }
+
+  private async ping(): Promise<boolean> {
+    try {
+      const pong = await this.cache.client.ping();
+      return pong === "PONG";
+    } catch (e) {
+      return false;
+    }
   }
 }
